+++ packages/ethereum/sdk/src/config/index.ts
@@ -6,14 +6,14 @@ import { polygonConfig } from "./polygon"
 import { devEthereumConfig } from "./dev"
 import { devPolygonConfig } from "./polygon-dev"
 import { testnetEthereumConfig } from "./testnet"
-import { stagingEthereumConfig } from "./staging"
-import { stagingPolygonConfig } from "./polygon-staging"
 import { mantleTestnetConfig } from "./testnet-mantle"
 import { mantleConfig } from "./mantle"
 import { arbitrumTestnetConfig } from "./testnet-arbitrum"
 import { arbitrumConfig } from "./arbitrum"
 import { zkSyncTestnetConfig } from "./testnet-zksync"
 import { zkSyncConfig } from "./zksync"
+import { chilizConfig } from "./chilliz"
+import { chilizTestnetConfig } from "./testnet-chiliz"
 
 export const configDictionary: Record<EthereumNetwork, EthereumConfig> = {
 	mainnet: mainnetConfig,
@@ -21,15 +21,15 @@ export const configDictionary: Record<EthereumNetwork, EthereumConfig> = {
 	polygon: polygonConfig,
 	"dev-ethereum": devEthereumConfig,
 	"dev-polygon": devPolygonConfig,
-	"staging-polygon": stagingPolygonConfig,
 	mantle: mantleConfig,
 	"testnet-mantle": mantleTestnetConfig,
 	testnet: testnetEthereumConfig,
-	staging: stagingEthereumConfig,
 	"testnet-arbitrum": arbitrumTestnetConfig,
-	"arbitrum": arbitrumConfig,
+	arbitrum: arbitrumConfig,
 	"testnet-zksync": zkSyncTestnetConfig,
-	"zksync": zkSyncConfig,
+	zksync: zkSyncConfig,
+	chiliz: chilizConfig,
+	"testnet-chiliz": chilizTestnetConfig,
 }
 
 export function getEthereumConfig(env: EthereumNetwork): EthereumConfig {