+++ packages/sdk-common/src/utils/blockchain.ts
@@ -1,13 +1,14 @@
 import { Blockchain } from "@rarible/api-client/build/models/Blockchain"
 
 export type EVMBlockchain = Blockchain.ETHEREUM | Blockchain.POLYGON | Blockchain.MANTLE
-| Blockchain.ARBITRUM | Blockchain.ZKSYNC
+| Blockchain.ARBITRUM | Blockchain.ZKSYNC | Blockchain.CHILIZ
 export const EVMBlockchains: EVMBlockchain[] = [
 	Blockchain.ETHEREUM,
 	Blockchain.POLYGON,
 	Blockchain.MANTLE,
 	Blockchain.ARBITRUM,
 	Blockchain.ZKSYNC,
+	Blockchain.CHILIZ,
 ]
 
 /**