+++ packages/sdk/jest.config.js
@@ -17,9 +17,9 @@ module.exports = {
 			outputDirectory: "reports",
 		}],
 	],
-    globals: {
-        "ts-jest": {
-            tsconfig: "tsconfig-build.json",
-        },
-    },
+	globals: {
+		"ts-jest": {
+			tsconfig: "tsconfig-build.json",
+		},
+	},
 }