+++ packages/sdk/src/config/index.ts
@@ -2,12 +2,10 @@ import { prodConfig } from "./prod"
 import type { RaribleSdkConfig, RaribleSdkEnvironment } from "./domain"
 import { developmentConfig } from "./development"
 import { testnetConfig } from "./testnet"
-import { stagingConfig } from "./staging"
 
 export const configsDictionary: Record<RaribleSdkEnvironment, RaribleSdkConfig> = {
 	development: developmentConfig,
 	testnet: testnetConfig,
-	staging: stagingConfig,
 	prod: prodConfig,
 }